# frontend/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Annotate Tool",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/stats": {
      "get": {
        "summary": "Project Stats",
        "operationId": "project_stats_api_v1_stats_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/tasks.csv": {
      "get": {
        "summary": "Task Report",
        "operationId": "task_report_api_v1_reports_tasks_csv_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-statistics/me": {
      "get": {
        "summary": "Personal Work Statistics",
        "operationId": "personal_work_statistics_api_v1_work_statistics_me_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "granularity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(day|week|month)$",
              "default": "day",
              "title": "Granularity"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalWorkStatisticsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-statistics/people": {
      "get": {
        "summary": "People Work Statistics",
        "operationId": "people_work_statistics_api_v1_work_statistics_people_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeopleWorkStatisticsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-statistics/me.csv": {
      "get": {
        "summary": "Personal Work Csv",
        "operationId": "personal_work_csv_api_v1_work_statistics_me_csv_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "granularity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(day|week|month)$",
              "default": "day",
              "title": "Granularity"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-statistics/people.csv": {
      "get": {
        "summary": "People Work Csv",
        "operationId": "people_work_csv_api_v1_work_statistics_people_csv_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PeopleWorkStatisticsOut": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/WorkMetricOut"
            },
            "type": "array",
            "title": "People"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "people"
        ],
        "title": "PeopleWorkStatisticsOut"
      },
      "PersonalWorkStatisticsOut": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "granularity": {
            "type": "string",
            "enum": [
              "day",
              "week",
              "month"
            ],
            "title": "Granularity"
          },
          "summary": {
            "$ref": "#/components/schemas/WorkMetricOut"
          },
          "periods": {
            "items": {
              "$ref": "#/components/schemas/WorkMetricOut"
            },
            "type": "array",
            "title": "Periods"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "granularity",
          "summary",
          "periods"
        ],
        "title": "PersonalWorkStatisticsOut"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "developer_admin",
          "annotation_manager",
          "reviewer",
          "annotator"
        ],
        "title": "Role"
      },
      "StatsOut": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Status"
          },
          "completion_rate": {
            "type": "number",
            "title": "Completion Rate"
          },
          "effective_video_seconds": {
            "type": "number",
            "title": "Effective Video Seconds"
          },
          "by_person": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "By Person"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "total",
          "by_status",
          "completion_rate",
          "effective_video_seconds",
          "by_person"
        ],
        "title": "StatsOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkMetricOut": {
        "properties": {
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "claimed_count": {
            "type": "integer",
            "title": "Claimed Count"
          },
          "first_submissions": {
            "type": "integer",
            "title": "First Submissions"
          },
          "resubmissions": {
            "type": "integer",
            "title": "Resubmissions"
          },
          "returned_count": {
            "type": "integer",
            "title": "Returned Count"
          },
          "final_approved_count": {
            "type": "integer",
            "title": "Final Approved Count"
          },
          "effective_video_seconds": {
            "type": "number",
            "title": "Effective Video Seconds"
          },
          "average_completion_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Completion Seconds"
          },
          "first_pass_rate": {
            "type": "number",
            "title": "First Pass Rate"
          },
          "rework_rate": {
            "type": "number",
            "title": "Rework Rate"
          },
          "review_claimed_count": {
            "type": "integer",
            "title": "Review Claimed Count"
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count"
          },
          "approved_count": {
            "type": "integer",
            "title": "Approved Count"
          },
          "rejected_count": {
            "type": "integer",
            "title": "Rejected Count"
          },
          "review_pass_rate": {
            "type": "number",
            "title": "Review Pass Rate"
          },
          "review_return_rate": {
            "type": "number",
            "title": "Review Return Rate"
          },
          "average_review_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Review Seconds"
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "period_end",
          "user_id",
          "display_name",
          "role",
          "claimed_count",
          "first_submissions",
          "resubmissions",
          "returned_count",
          "final_approved_count",
          "effective_video_seconds",
          "average_completion_seconds",
          "first_pass_rate",
          "rework_rate",
          "review_claimed_count",
          "review_count",
          "approved_count",
          "rejected_count",
          "review_pass_rate",
          "review_return_rate",
          "average_review_seconds"
        ],
        "title": "WorkMetricOut"
      }
    }
  }
}